refactor: batchUpdate, запрос друзей через БД и лайки

+++ src/main/java/ru/yandex/practicum/filmorate/service/FriendService.java
@@ -7,10 +7,7 @@ import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.friend.FriendStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @Service
 public class FriendService {
@@ -25,26 +22,26 @@ public class FriendService {
     }
 
     public User addFriend(long id, long friendId) {
-        friendStorage.addFriend(id, friendId, Status.CONFIRMED);
-        return userStorage.get(id);
+        if (userStorage.contains(id) && userStorage.contains(friendId)) {
+            friendStorage.addFriend(id, friendId, Status.CONFIRMED);
+            return userStorage.get(id);
+        }
+        return null;
     }
 
     public User deleteFriend(long id, long friendId) {
-        friendStorage.deleteFriend(id, friendId);
-        return userStorage.get(id);
+        if (userStorage.contains(id) && userStorage.contains(friendId)) {
+            friendStorage.deleteFriend(id, friendId);
+            return userStorage.get(id);
+        }
+        return null;
     }
 
     public List<User> getFriends(long id) {
-        Set<Long> friendsId = friendStorage.getFriends(id);
-        return friendsId.stream().map(userStorage::get).collect(Collectors.toList());
+        return friendStorage.getFriends(id);
     }
 
     public List<User> getCommonFriends(long id, long otherId) {
-        Set<Long> intersection = new HashSet<>(friendStorage.getFriends(id));
-        intersection.retainAll(friendStorage.getFriends(otherId));
-        return intersection
-                .stream()
-                .map(userStorage::get)
-                .collect(Collectors.toList());
+        return friendStorage.getCommonFriends(id, otherId);
     }
 }
